Add the '-e' / '--no-err-files' option.
In some cases TESTDIR is read-only, so storing an .err file on a test
failure is impossible which make cram itself fail.

For example, if Cram is invoked as a part of 'make distcheck', which
deliberately runs a VPATH build with 'srcdir' and all its contents made
read-only, any test failure leads to Cram failing with a Python
traceback in the test log. Having a diff file instead would still be
handy to diagnose the failure.

With the new '-e' / '--no-err-files' option Cram does not try to create
.err files on test failures, but still prints the diff.