Restrict the scope of the test selection aimed for the code coverage of prysk
In order to be self sustained, it is important that the prysk tests by themselfs provide
the expected code- and test- coverage for the prysk package.
Relying on tests for other scripts/tools will impose implicit dependencies regarding testing.
Such test depenencies are not desired and therefore the coverage target
was configured differently.